additional images and adjustment to converter code locations
